Final Changes
I completed the page. Unfortunately, I'm disappointed that I couldn't
get the mic in the search bar due to my lack of knowledge. (I'm a
perfectionist) Anyways, I added the footer and cleaned up the css and
html. Will be minifying the css using grunt next to complete. I will be
coming back to this and adding the mic at a later date because it's
bothering me but I've spent too many hours trying to figure it out for a
simple project.